refactor(core): return persistent_kernel.py toward fork; DSv3 wrappers move to models/deepseek_v3/tasks.py
User decision: the core file must stay close to upstream; model-
specific registration helpers live with the model. Option-1 move:

- 31 DSv3-only wrapper methods leave PersistentKernel verbatim and
  become module functions in models/deepseek_v3/tasks.py (pk-first
  arg); builder + 14 tests rewired (22 + 39 call sites, bind-checked).
- the fork-named MLA publics return (tp1 byte-identical; tp2/4/8 keep
  the campaign grid-math bodies); run_test_mode and the catalog-ABI
  mla_decode_layer are restored fork-verbatim.
- pk diff vs fork (de8745d9): 2573 → 1132 lines; the remainder is
  the user-sanctioned nvcc region (rdc=true unconditional for
  multi-GPU), tensor_view, load-bearing infra (HARD_CODE ABI,
  attach_input views, profiler CSV), and 21 fork-present methods
  whose body fixes have live C++ counterparts — listed in the
  restructuring report for per-item review.
- registered task graph unchanged by construction (verbatim bodies,
  same task names/params).
