+++ tests/runtime_python/blackwell/sm100_linear_fp8_bmm/test_linear_fp8_bmm_testmode.py
@@ -27,6 +27,7 @@ sys.path.insert(0, os.path.join(
 
 import mirage  # noqa: E402
 from mirage.mpk.persistent_kernel import PersistentKernel  # noqa: E402
+from mirage.mpk.models.deepseek_v3 import tasks as dsv3_tasks
 from sm100_fp8_scale_layout import (  # noqa: E402
     quantize_to_fp8_packed_ue8m0,
     dequant_from_packed_ue8m0,
@@ -190,7 +191,7 @@ def _run_case(label: str, batch: int, num_heads: int, d_in: int,
     w_sc = pk.attach_input(weight_scale, name="bmm_weight_scale")
     o = pk.attach_input(output, name="bmm_output")
 
-    pk.linear_fp8_bmm_layer(dense=False,
+    dsv3_tasks.linear_fp8_bmm_layer(pk, dense=False,
         
         input_fp8=i_fp8, input_scale=i_sc,
         weight_fp8=w_fp8, weight_scale=w_sc,
